# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /districts:
    get:
      operationId: listDistricts
      tags: [districts]
      summary: List districts
      parameters:
        - name: stateId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of districts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/District"
    post:
      operationId: createDistrict
      tags: [districts]
      summary: Create a district
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDistrictBody"
      responses:
        "201":
          description: Created district
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/District"
  /institutions:
    get:
      operationId: listInstitutions
      tags: [institutions]
      summary: List institutions
      parameters:
        - name: districtId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of institutions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Institution"
    post:
      operationId: createInstitution
      tags: [institutions]
      summary: Create an institution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstitutionBody"
      responses:
        "201":
          description: Created institution
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Institution"
...
  /exams/start:
    post:
      operationId: startExam
      tags: [exams]
      summary: Start an exam attempt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartExamBody"
      responses:
        "201":
          description: Exam attempt started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamAttempt"
  /exams/submit:
    post:
      operationId: submitExam
      tags: [exams]
      summary: Submit exam answers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitExamBody"
      responses:
        "200":
          description: Exam result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /exams/attempts:
    get:
      operationId: listExamAttempts
      tags: [exams]
      summary: List exam attempts for current user
      parameters:
        - name: quizId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamAttempt"
...
components:
  schemas:
...
    District:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        stateId:
          type: integer
        stateName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - stateId
        - createdAt
    CreateDistrictBody:
      type: object
      properties:
        name:
          type: string
        stateId:
          type: integer
      required:
        - name
        - stateId
    Institution:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        districtId:
          type: integer
        districtName:
          type: string
        stateName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - districtId
        - createdAt
    CreateInstitutionBody:
      type: object
      properties:
        name:
          type: string
        districtId:
          type: integer
      required:
        - name
        - districtId
...
    ExamAttempt:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: string
        startedAt:
          type: string
          format: date-time
        submittedAt:
          type: ["string", "null"]
          format: date-time
        score:
          type: ["integer", "null"]
        totalQuestions:
          type: integer
        correctAnswers:
          type: ["integer", "null"]
        status:
          type: string
          enum: [IN_PROGRESS, SUBMITTED, TIMED_OUT]
      required:
        - id
        - userId
        - quizId
        - startedAt
        - totalQuestions
        - status
    StartExamBody:
      type: object
      properties:
        quizId:
          type: integer
      required:
        - quizId
    SubmitExamBody:
      type: object
      properties:
        attemptId:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerItem"
      required:
        - attemptId
        - answers
    AnswerItem:
      type: object
      properties:
        questionId:
          type: integer
        selectedOption:
          type: integer
      required:
        - questionId
        - selectedOption
    ExamResult:
      type: object
      properties:
        attemptId:
          type: integer
        score:
          type: integer
        totalQuestions:
          type: integer
        correctAnswers:
          type: integer
        percentage:
          type: number
      required:
        - attemptId
        - score
        - totalQuestions
        - correctAnswers
        - percentage
...